+++ examples/sort lines.py
@@ -1,0 +1,6 @@
+import sys
+
+lines = sys.argv[1].split("\n")
+lines = sorted(lines)
+for line in lines:
+    print(line)