+++ CHANGELOG.md
@@ -3,12 +3,18 @@
 https://keepachangelog.com
 
 ## [1.0.0] - 2021-10-23
-
 - Initial release
 ## [1.0.1]
-
 - Fixed main in package.json
-
 ## [1.0.2]
 
 - Added project icon
+## [1.0.3] - 2021-10-25
+### Changed
+- Documentation
+- `tops.tops` command is now `tops.applyOperation`
+- `SortLines.py` example is simplified
+### Added
+- Additional examples
+- Configurable property `tops.pythonExecutablePath`
+- `Ctrl-\` keybinding for running `tops.applyOperation`